Removed interface MalcolmModel and created AbstractMalcolmModel
http://jira.diamond.ac.uk/browse/DAQ-358
