+++ org.eclipse.scanning.example/src/org/eclipse/scanning/example/malcolm/TestMalcolmModel.java
@@ -2,8 +2,7 @@ package org.eclipse.scanning.example.malcolm;
 
 import org.eclipse.scanning.api.annotation.ui.FieldDescriptor;
 import org.eclipse.scanning.api.annotation.ui.FileType;
-import org.eclipse.scanning.api.device.models.AbstractDetectorModel;
-import org.eclipse.scanning.api.points.IPointGenerator;
+import org.eclipse.scanning.api.device.models.AbstractMalcolmModel;
 
 /**
  * A Test Malcolm Model.
@@ -11,16 +10,11 @@ import org.eclipse.scanning.api.points.IPointGenerator;
  * @author Matt Taylor
  *
  */
-public class TestMalcolmModel extends AbstractDetectorModel {
+public class TestMalcolmModel extends AbstractMalcolmModel {
 
 	@FieldDescriptor(label="Directory", hint="The directory where the h5 files will be created", file=FileType.NEW_FOLDER)
 	private String filePath;
 	
-	/**
-	 * TODO Why are generators in the malcolm model?
-	 */
-	@FieldDescriptor(visible=false)
-	private IPointGenerator<?> generator;
 
 	public String getFilePath() {
 		return filePath;
@@ -30,21 +24,11 @@ public class TestMalcolmModel extends AbstractDetectorModel {
 		this.filePath = filePath;
 	}
 
-	public IPointGenerator<?> getGenerator() {
-		return generator;
-	}
-
-	public void setGenerator(IPointGenerator<?> generator) {
-		this.generator = generator;
-	}
-	
-
 	@Override
 	public int hashCode() {
 		final int prime = 31;
 		int result = super.hashCode();
 		result = prime * result + ((filePath == null) ? 0 : filePath.hashCode());
-		result = prime * result + ((generator == null) ? 0 : generator.hashCode());
 		return result;
 	}
 
@@ -61,11 +45,6 @@ public class TestMalcolmModel extends AbstractDetectorModel {
 			if (other.filePath != null)
 				return false;
 		} else if (!filePath.equals(other.filePath))
-			return false;
-		if (generator == null) {
-			if (other.generator != null)
-				return false;
-		} else if (!generator.equals(other.generator))
 			return false;
 		return true;
 	}
